app/requirements.txt
Cython
sklearn
inflection
xgboost
Flask
numpy
pandas
requests
plotly-express
streamlit